Add diagnostic and audit log v2 UI
