+++ apps/console/src/extensions/components/logs/components/advanced-filter-popup.tsx
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2022, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2022, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -26,7 +26,7 @@ import React, {
 } from "react";
 import { useTranslation } from "react-i18next";
 import { Divider, Form, Grid } from "semantic-ui-react";
-import { InterfaceLogsFilter } from "../models/diagnostic-log-models";
+import { InterfaceLogsFilter } from "../models/log-models";
 
 /**
  * Filter attribute field identifier.
