+++ apps/console/src/extensions/components/logs/hooks/use-fetch.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2022, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2022, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,49 +16,21 @@
  * under the License.
  */
 
-import { useCallback, useEffect, useState } from "react";
-import { getDiagnosticLogs } from "../api";
-import { InterfaceDiagnosticLogsRequest, InterfaceDiagnosticLogsResponse } from "../models";
+import { useLogs } from "../api/useLogs";
+import { InterfaceLogEntry, InterfaceLogsRequest } from "../models/log-models";
 
 /**
  * Custom hook for fetching logs
  * @param requestPayload - logs search API endpoint POST request payload
  * @returns error, list, loading, next, previous
  */
-function useFetch(requestPayload: InterfaceDiagnosticLogsRequest): any {
-    const [ loading, setLoading ] = useState(true);
-    const [ error, setError ] = useState(false);
-    const [ list, setList ] = useState([]);
-    const [ next, setNext ] = useState(null);
-    const [ previous, setPrevious ] = useState(null);
+function useFetch(requestPayload: InterfaceLogsRequest): any {
+    const { data, error, isLoading } = useLogs(requestPayload);
+    const list: InterfaceLogEntry[] = data?.logs || [];
+    const next: string | null = data?.nextToken || null;
+    const previous: string | null = data?.previousToken || null;
 
-    const sendQuery: () => Promise<void> = useCallback(async () => {
-        try {
-            setLoading(true);
-            setError(false);
-            const data: InterfaceDiagnosticLogsResponse = await getDiagnosticLogs(requestPayload);
-
-            setList(data.logs);
-            setNext(data.nextToken);
-            setPrevious(data.previousToken);
-            setLoading(false);
-        } catch (err) {
-            setError(err);
-        }
-    }, [ requestPayload ]);
-
-    useEffect(() => {
-        if(!requestPayload ||
-            (!requestPayload.endTime && !requestPayload.nextToken && !requestPayload.previousToken) ||
-            (!requestPayload.startTime&& !requestPayload.nextToken && !requestPayload.previousToken)
-        )
-        {
-            return;
-        }
-        sendQuery();
-    }, [ requestPayload ]);
-
-    return { error, list, loading, next, previous };
+    return { error, list, loading: isLoading, next, previous };
 }
 
 export default useFetch;
