+++ apps/console/src/extensions/i18n/resources/si-LK/extensions.ts
@@ -2686,16 +2686,16 @@ export const extensions: Extensions = {
                                     progressiveEnrollment: {
                                         description: "පුරනය වීමේදී මුර යතුර සඳහා ලියාපදිංචි වීමට පරිශීලකයින්ට ඉඩ දීමට මෙම විකල්පය සක්‍රිය කරන්න.",
                                         label: "ප්‍රගතිශීලී මුර යතුර ලියාපදිංචිය:",
-                                        note: "මුර යතුර පළමු සාධකය ලෙස සකසා ඇත්නම්, යෙදුමේ <3>පිවිසීමේ ක්‍රමය</3> " + 
+                                        note: "මුර යතුර පළමු සාධකය ලෙස සකසා ඇත්නම්, යෙදුමේ <3>පිවිසීමේ ක්‍රමය</3> " +
                                         "ටැබය යටතේ පහත <1> අනුවර්තන ස්ක්‍රිප්ට්</1> එක් කළ යුතුය. " +
                                         "මෙම ස්ක්‍රිප්ටය සැකිල්ල මත පදනම් වූ මුරපද සැකසුම සමඟ ස්වයංක්‍රීයව " +
                                         "එකතු වන අතර මුරපද ලියාපදිංචි කිරීමට පෙර පරිශීලකයාගේ අනන්‍යතාවය " +
-                                        "තහවුරු කිරීමට භාවිතා කරයි. කෙසේ වෙතත්, ඔබ අච්චුවක් නොමැතිව මුරපද " + 
+                                        "තහවුරු කිරීමට භාවිතා කරයි. කෙසේ වෙතත්, ඔබ අච්චුවක් නොමැතිව මුරපද " +
                                         "වින්‍යාස කරන්නේ නම්, ස්ක්‍රිප්ට් එක හස්තීයව එක් කිරීමට මතක තබා ගන්න."
                                     },
                                     usernamelessAuthentication: {
-                                        description: "මෙම විශේෂාංගය සක්‍රීය කිරීමෙන් පරිශීලකයින්ට පරිශීලක " + 
-                                        "නාමය ඇතුළත් නොකර මුරපදයක් සමඟින් පුරනය වීමට ඉඩ සලසයි, එය වඩාත් " + 
+                                        description: "මෙම විශේෂාංගය සක්‍රීය කිරීමෙන් පරිශීලකයින්ට පරිශීලක " +
+                                        "නාමය ඇතුළත් නොකර මුරපදයක් සමඟින් පුරනය වීමට ඉඩ සලසයි, එය වඩාත් " +
                                         "විධිමත් පුරනය වීමේ අත්දැකීමක් නිර්මාණය කරයි.",
                                         label: "පරිශීලක නාම රහිත සත්‍යාපනය:"
                                     }
@@ -2794,7 +2794,8 @@ export const extensions: Extensions = {
                     }
                 },
                 searchBar: {
-                    placeholder: "ට්‍රේස් හැඳුනුම්පත, ක්‍රියා හැඳුනුම්පත, සේවාලාභී හැඳුනුම්පත, ප්‍රතිඵල පණිවිඩය හෝ ප්‍රතිඵල තත්ත්වය අනුව ලොග් සොයන්න"
+                    placeholderDiagnostic: "ට්‍රේස් හැඳුනුම්පත, ක්‍රියා හැඳුනුම්පත, සේවාලාභී හැඳුනුම්පත, ප්‍රතිඵල පණිවිඩය හෝ ප්‍රතිඵල තත්ත්වය අනුව ලොග් සොයන්න",
+                    placeholderAudit : "ක්‍රියාව, ඉලක්ක හැඳුනුම්පත, ආරම්භක හැඳුනුම්පත, ඉල්ලීම් හැඳුනුම්පත අනුව ලොග සොයන්න"
                 },
                 refreshMessage: {
                     text: "අවසන් වරට ලොග ලබා ගත්තේ ",
@@ -2805,6 +2806,9 @@ export const extensions: Extensions = {
                 },
                 queryButton: {
                     label: "ලොග ලබා ගන්න"
+                },
+                downloadButton : {
+                    label : "ලොග් දත්ත බාගන්න"
                 },
                 delayMessage: {
                     text: "සමහර විමසුම් පූරණය වීමට වැඩි කාලයක් ගත විය හැක."
@@ -2818,7 +2822,7 @@ export const extensions: Extensions = {
             notifications: {
                 genericError: {
                     subtitle: {
-                        0: "විනිශ්චය ලොග ලබා ගැනීමට නොහැකි විය.",
+                        0: "ලොග ලබා ගැනීමට නොහැකි විය.",
                         1: "කරුණාකර නැවත උත්සාහ කරන්න."
                     },
                     title: "යමක් වැරදී ඇත"
@@ -2848,8 +2852,8 @@ export const extensions: Extensions = {
                 }
             },
             pageHeader: {
-                description: "නිෂ්පාදන යෙදුම් ගැටළු නිරාකරණය කිරීමට ඔබගේ ලඝු-සටහන් විමසන්න.",
-                title: "විනිශ්චය ලොග්"
+                description: "ගැටළු නිරාකරණය කිරීමට සහ සම්පත් ක්‍රියාකාරකම් නිරීක්ෂණය කිරීමට ඔබගේ ලඝු-සටහන් විමසන්න.",
+                title: "ලොග්"
             },
             tooltips: {
                 copy: "ක්ලිප් පුවරුවට පිටපත් කරන්න"
